Fix[u_distortion]: fix multiplication warning

+++ src/xrt/auxiliary/util/u_distortion.c
@@ -82,9 +82,9 @@ u_distortion_cardboard_calculate(const struct u_cardboard_distortion_arguments *
 		// clang-format on
 
 		// Turn into tanangles
-		values->screen.size.x /= args->screen_to_lens_distance_meters * view_count;
+		values->screen.size.x /= args->screen_to_lens_distance_meters * (float) view_count;
 		values->screen.size.y /= args->screen_to_lens_distance_meters;
-		values->screen.offset.x /= args->screen_to_lens_distance_meters * view_count;
+		values->screen.offset.x /= args->screen_to_lens_distance_meters * (float) view_count;
 		values->screen.offset.y /= args->screen_to_lens_distance_meters;
 
 		// Tanangle to texture coordinates
@@ -93,9 +93,6 @@ u_distortion_cardboard_calculate(const struct u_cardboard_distortion_arguments *
 		values->texture.offset.x = tanf(-args->fov.angle_left);
 		values->texture.offset.y = tanf(-args->fov.angle_down);
 
-		// Fix up views not covering the entire screen.
-		//values->screen.size.x /= view_count;
-		//values->screen.offset.x -= values->screen.size.x * i;
 	}
 }
